test: 혹시 AWS SDK가 문제였을까?
