+++ src/test/java/dev/sijunyang/celog/surpport/ncp/ImageUploadServiceImplIntegrationTest.java
@@ -1,68 +1,68 @@
-package dev.sijunyang.celog.surpport.ncp;
-
-import dev.sijunyang.celog.core.global.error.nextVer.InvalidInputException;
-import io.awspring.cloud.s3.S3Template;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.core.io.FileSystemResource;
-import org.springframework.core.io.Resource;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.util.StringUtils;
-
-import java.nio.charset.StandardCharsets;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
-
-@ActiveProfiles("test")
-@SpringBootTest
-class ImageUploadServiceImplIntegrationTest {
-
-    @Autowired
-    private ImageUploadServiceImpl service;
-
-    @Autowired
-    private S3Template template;
-
-    @Autowired
-    private NcpProperties ncpProperties;
-
-    // @Test
-    // void shouldUploadImage() {
-    // // Given
-    // String fileLocation =
-    // "./src/test/java/dev/sijunyang/celog/surpport/ncp/test-image.jpg";
-    // Resource resource = new FileSystemResource(fileLocation);
-    //
-    // // When
-    // String result = this.service.execute(resource);
-    //
-    // // Then
-    // assertNotNull(result);
-    //
-    // // 테스트가 성공했으므로 필요없는 데이터는 삭제
-    // String filename = StringUtils.getFilename(StringUtils.uriDecode(result,
-    // StandardCharsets.UTF_8));
-    // template.deleteObject(ncpProperties.bucketName(), filename);
-    // }
-    //
-    // @Test
-    // void shouldThrowInvalidInputExceptionWhenFileExtensionIsNotSupported() {
-    // // Given
-    // String fileLocation =
-    // "./src/test/java/dev/sijunyang/celog/surpport/ncp/test-text.txt";
-    // Resource resource = new FileSystemResource(fileLocation);
-    //
-    // // When & Then
-    // assertThrows(InvalidInputException.class, () -> service.execute(resource));
-    // }
-
-    @Test
-    void shouldGetNcpProperties() {
-        assertNotNull(ncpProperties);
-        assertNotNull(ncpProperties.bucketName());
-        assertNotNull(ncpProperties.region());
-    }
-
-}
+// package dev.sijunyang.celog.surpport.ncp;
+//
+// import dev.sijunyang.celog.core.global.error.nextVer.InvalidInputException;
+// import io.awspring.cloud.s3.S3Template;
+// import org.junit.jupiter.api.Test;
+// import org.springframework.beans.factory.annotation.Autowired;
+// import org.springframework.boot.test.context.SpringBootTest;
+// import org.springframework.core.io.FileSystemResource;
+// import org.springframework.core.io.Resource;
+// import org.springframework.test.context.ActiveProfiles;
+// import org.springframework.util.StringUtils;
+//
+// import java.nio.charset.StandardCharsets;
+//
+// import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+// import static org.junit.jupiter.api.Assertions.*;
+//
+// @ActiveProfiles("test")
+// @SpringBootTest
+// class ImageUploadServiceImplIntegrationTest {
+//
+// @Autowired
+// private ImageUploadServiceImpl service;
+//
+// @Autowired
+// private S3Template template;
+//
+// @Autowired
+// private NcpProperties ncpProperties;
+//
+// // @Test
+// // void shouldUploadImage() {
+// // // Given
+// // String fileLocation =
+// // "./src/test/java/dev/sijunyang/celog/surpport/ncp/test-image.jpg";
+// // Resource resource = new FileSystemResource(fileLocation);
+// //
+// // // When
+// // String result = this.service.execute(resource);
+// //
+// // // Then
+// // assertNotNull(result);
+// //
+// // // 테스트가 성공했으므로 필요없는 데이터는 삭제
+// // String filename = StringUtils.getFilename(StringUtils.uriDecode(result,
+// // StandardCharsets.UTF_8));
+// // template.deleteObject(ncpProperties.bucketName(), filename);
+// // }
+// //
+// // @Test
+// // void shouldThrowInvalidInputExceptionWhenFileExtensionIsNotSupported() {
+// // // Given
+// // String fileLocation =
+// // "./src/test/java/dev/sijunyang/celog/surpport/ncp/test-text.txt";
+// // Resource resource = new FileSystemResource(fileLocation);
+// //
+// // // When & Then
+// // assertThrows(InvalidInputException.class, () -> service.execute(resource));
+// // }
+//
+// @Test
+// void shouldGetNcpProperties() {
+// assertNotNull(ncpProperties);
+// assertNotNull(ncpProperties.bucketName());
+// assertNotNull(ncpProperties.region());
+// }
+//
+// }
